move switch_client related code out of main file
